Move Auth0 logic to Auth component.
