+++ frontend/src/components/Auth.js
@@ -1,9 +1,22 @@
 import React, { Component } from 'react';
+import auth0 from 'auth0-js';
+import { Auth0Lock } from 'auth0-lock';
 import styled from 'styled-components';
+import { connect } from 'react-redux';
 import { Link } from 'react-router-dom';
 import LoginDropdown from './LoginDropdown.js';
 import chevron from '../assets/img/chevron.png';
 // import PropTypes from 'prop-types';
+
+// globals
+import {
+  auth0Domain,
+  auth0ClientID,
+  auth0RedirectUri
+} from '../globals/globals.js';
+
+// action creators
+import { auth0Login, displayError } from '../store/actions/index.js';
 
 /***************************************************************************************************
  ********************************************** Styles *********************************************
@@ -56,10 +69,53 @@ const Login = styled.a`
 class Auth extends Component {
   constructor(props) {
     super(props);
+    this.webAuth.parseHash((err, authResult) => {
+      if (authResult) {
+        const { accessToken, expiresIn } = authResult;
+        const expiresAt = JSON.stringify(
+          expiresIn * 1000 + new Date().getTime()
+        );
+        localStorage.setItem('symposium_auth0_access_token', accessToken);
+        localStorage.setItem('symposium_auth0_expires_at', expiresAt);
+        return this.props.auth0Login(accessToken);
+      } else if (err) this.props.displayError(err);
+    });
+
     this.state = {
       isLoginDropdownClicked: false,
       setIsLoginDropdownClicked: false
     };
+    
+  }
+
+  authLockOptions = {
+    rememberLastLogin: false
+  };
+
+  lock = new Auth0Lock(auth0ClientID, auth0Domain, this.authLockOptions);
+
+  webAuth = new auth0.WebAuth({
+    domain: auth0Domain,
+    clientID: auth0ClientID,
+    redirectUri: auth0RedirectUri
+  });
+
+  handleAuth0Login = () => {
+    this.lock.show();
+  };
+
+  componentDidUpdate(prevProps) {
+    if (this.props.location.pathname !== prevProps.location.pathname) {
+      this.authLockOptions = {
+        rememberLastLogin: false
+      };
+      this.lock = new Auth0Lock(auth0ClientID, auth0Domain, this.authLockOptions);
+      this.webAuth = new auth0.WebAuth({
+        domain: auth0Domain,
+        clientID: auth0ClientID,
+        redirectUri: auth0RedirectUri
+      });
+    }
   }
 
   toggleLoginDropdown = ev => {
@@ -79,7 +135,7 @@ class Auth extends Component {
           You are not authorized to view this content, please click below to
           LOGIN.
         </p>
-        <button onClick={this.props.handleAuth0Login}>Login via Auth0</button>
+        <button onClick={this.handleAuth0Login}>Login via Auth0</button>
 
         <DivAuth>
           <LinkRegister to='/register'>Register</LinkRegister> |{' '}
@@ -107,4 +163,4 @@ class Auth extends Component {
 //   propertyName: PropTypes.string
 // }
 
-export default Auth;
+export default connect(null, { auth0Login, displayError })(Auth);
